requirements.txt
numpy
tensorflow
pyyaml
pandas
tables
keras
streamlit
joblib
Pillow